refactor: split Backstage error type into separate file
