+++ plugins/backstage-plugin-coder/src/components/CoderProvider/CoderAuthProvider.tsx
@@ -11,9 +11,8 @@ import {
   useQuery,
   useQueryClient,
 } from '@tanstack/react-query';
-
+import { BackstageHttpError } from '../../api/errors';
 import {
-  BackstageHttpError,
   CODER_QUERY_KEY_PREFIX,
   authQueryKey,
   authValidation,
@@ -158,7 +157,7 @@ export const CoderAuthProvider = ({ children }: CoderAuthProviderProps) => {
       const queryError = event.query.state.error;
       const shouldRevalidate =
         !isRefetchingTokenQuery &&
-        queryError instanceof BackstageHttpError &&
+        BackstageHttpError.isInstance(queryError) &&
         queryError.status === 401;
 
       if (!shouldRevalidate) {
@@ -240,7 +239,7 @@ function generateAuthState({
     };
   }
 
-  if (authValidityQuery.error instanceof BackstageHttpError) {
+  if (BackstageHttpError.isInstance(authValidityQuery.error)) {
     const deploymentLikelyUnavailable =
       authValidityQuery.error.status === 504 ||
       (authValidityQuery.error.status === 200 &&
